Configured props for node information to move between components
Information (with manual input at this stage) can be transferred from the dashboard component to the drop down node list using props. The manual input needs to be updated to the nodes that the user selects.
