+++ src/components/DataTable/ExpandableCard.js
@@ -1,25 +1,6 @@
-import React from 'react';
+import React, { PropTypes } from 'react'
 import {Card, CardActions, CardHeader, CardText} from 'material-ui/Card';
 import FlatButton from 'material-ui/FlatButton';
-
-
-const nodeMap = [
-    {
-         title: 'Node 263',
-         minPower: 'Min Power: 288 Watts',
-         maxPower: 'Max Power: 465 Watts'
-        },
-       {
-         title: 'Node 5',
-         minPower: 'Min Power: 123 Watts',
-         maxPower: 'Max Power: 354 Watts'     
-       },
-        {
-         title: 'Node 135',
-         minPower: 'Min Power: 7 Watts',
-         maxPower: 'Max Power: 45 Watts'       
-       }
-     ] ;
   
 export default class ExpandableCard extends React.Component {
     constructor(props) {
@@ -29,7 +10,7 @@ export default class ExpandableCard extends React.Component {
     }
 
     render() {
-        const nodeCards= nodeMap.map(p => 
+        const nodeCards= this.props.nodeMap.map(p => 
         <Card>
             <CardHeader
               title={p.title}
@@ -41,6 +22,22 @@ export default class ExpandableCard extends React.Component {
               <p>{p.maxPower}</p>
             </CardText>
         </Card>)
+console.log(this.props);
+/*const nodeCards = nodeMap.map(function(object, i){
+  return <Card key = {i}>
+  <CardHeader
+    title={object.title}
+    actAsExpander={true}
+    showExpandableButton={true}
+    key = {i+1}
+  />
+  <CardText expandable={true} key={i+2}>
+    <p>{object.maxPower}</p>
+    <p>{object.maxPower}</p>
+  </CardText>
+</Card>});
+*/
+
 
         return (
             <div> 
@@ -49,3 +46,7 @@ export default class ExpandableCard extends React.Component {
         );
     }
   }
+
+
+
+
